pyproject.toml
[project]
name = "maxentsolver"
version = "0.1.0"
description = "A Python package for fitting maximum entropy models to BOLD time-series data. The toolkit automatically binarizes fMRI/BOLD signals, constructs pairwise maximum entropy (MaxEnt) models, and provides utilities to analyze the inferred interaction network."
authors = [
    {name = "Marc Vöhringer Carrera"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytorch (>=1.0.2,<2.0.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
